Split upstream's CSV conventions out of the convert stage
convert owned two things: how MeteoSwiss writes a CSV, and Bronze ->
Parquet. Five modules imported it and four wanted only the first —
readers for the load path, transfer for one byte-level helper, api, and
the Databricks script, which writes Delta and never produces a Parquet
file. So the download engine depended on the Parquet converter, and the
gridded read path did too, through transfer.

meteocsv holds the conventions: the encoding fallback, the dtype
overrides, the KEY;VALUE preamble, the filename rules. convert keeps the
stage and is now imported by registry alone. The polars dtype-error
pattern was matched in both halves and is one function in meteocsv now.

test_layering asserts the graph rather than a diagram: acyclic, convert
has one consumer, meteocsv sits under everything that reads a file.

Also: exists was defined identically in client and grids, and
already_current only in client. Both are SkipRules, so both live in
transfer next to the type that names them. The climate-scenario CSV
fixture was copied in two test modules; it is in conftest now.
